Datenbankzugirffe aus dem ViewController in eigenen Service
